we create a new 12 column grid within the enclosing span8 and split it into three span4's. Then we style them in such
a way that the big number is up top in an h3, and the small descriptor is underneath in a p. Finally, we use linediv-l and linediv-c in an interesting way with the media queries, to give vertical lines when these spans are laid out horizontally (in a desktop view) and horizontal lines when laid out vertically (in a mobile view).
https://developer.mozilla.org/en-US/docs/Web/CSS/border

For the social buttons we need to use bootstrap-socialbuttons and this forces a FontAwesome import as well (both of which are in the <head>).
http://noizwaves.github.io/bootstrap-socialbuttons